+++ current_trades1.csv
@@ -1,0 +1,7 @@
+2020-01-01,PEAR,BUY,10000,1011.0
+2020-09-12,PEAR,SELL,5000,6713.45
+2020-10-30,PEAR,BUY,1000,785.3
+2021-01-02,PEAR,SELL,5500,16891.2
+2005-04-05,IBN,BUY,500,7820.5
+2020-09-12,IBN,SELL,500,4319.32
+2000-07-13,XYZ,BUY,333,99100.42
